refactor(dto): move DTOs to a separate package with package-info
