+++ src/test/java/com/conveyal/gtfs/dto/FrequencyDTO.java
@@ -1,4 +1,4 @@
-package com.conveyal.gtfs.util;
+package com.conveyal.gtfs.dto;
 
 public class FrequencyDTO {
     public String trip_id;
