+++ src/test/java/com/conveyal/gtfs/dto/RouteDTO.java
@@ -1,4 +1,4 @@
-package com.conveyal.gtfs.util;
+package com.conveyal.gtfs.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
