+++ src/test/java/com/conveyal/gtfs/dto/ShapePointDTO.java
@@ -1,4 +1,4 @@
-package com.conveyal.gtfs.util;
+package com.conveyal.gtfs.dto;
 
 // TODO add fields
 public class ShapePointDTO {
