+++ src/test/java/com/conveyal/gtfs/dto/TripDTO.java
@@ -1,4 +1,4 @@
-package com.conveyal.gtfs.util;
+package com.conveyal.gtfs.dto;
 
 public class TripDTO {
     public Integer id;
