+++ src/test/java/com/conveyal/gtfs/loader/JDBCTableWriterTest.java
@@ -1,19 +1,19 @@
 package com.conveyal.gtfs.loader;
 
 import com.conveyal.gtfs.TestUtils;
-import com.conveyal.gtfs.util.CalendarDTO;
-import com.conveyal.gtfs.util.FareDTO;
-import com.conveyal.gtfs.util.FareRuleDTO;
-import com.conveyal.gtfs.util.FeedInfoDTO;
-import com.conveyal.gtfs.util.FrequencyDTO;
+import com.conveyal.gtfs.dto.CalendarDTO;
+import com.conveyal.gtfs.dto.FareDTO;
+import com.conveyal.gtfs.dto.FareRuleDTO;
+import com.conveyal.gtfs.dto.FeedInfoDTO;
+import com.conveyal.gtfs.dto.FrequencyDTO;
 import com.conveyal.gtfs.util.InvalidNamespaceException;
-import com.conveyal.gtfs.util.PatternDTO;
-import com.conveyal.gtfs.util.PatternStopDTO;
-import com.conveyal.gtfs.util.RouteDTO;
-import com.conveyal.gtfs.util.ShapePointDTO;
-import com.conveyal.gtfs.util.StopDTO;
-import com.conveyal.gtfs.util.StopTimeDTO;
-import com.conveyal.gtfs.util.TripDTO;
+import com.conveyal.gtfs.dto.PatternDTO;
+import com.conveyal.gtfs.dto.PatternStopDTO;
+import com.conveyal.gtfs.dto.RouteDTO;
+import com.conveyal.gtfs.dto.ShapePointDTO;
+import com.conveyal.gtfs.dto.StopDTO;
+import com.conveyal.gtfs.dto.StopTimeDTO;
+import com.conveyal.gtfs.dto.TripDTO;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
